Split changes for pr(10)
